refactor(process): unify error handling with InvalidDataException and update DataError types
