+++ core/src/main/java/kafka/automq/table/deserializer/proto/HeaderBasedSchemaResolutionResolver.java
@@ -21,6 +21,7 @@ package kafka.automq.table.deserializer.proto;
 
 import kafka.automq.table.deserializer.SchemaResolutionResolver;
 import kafka.automq.table.deserializer.proto.schema.MessageIndexes;
+import kafka.automq.table.process.exception.InvalidDataException;
 
 import org.apache.kafka.common.errors.SerializationException;
 import org.apache.kafka.common.record.Record;
@@ -65,13 +66,13 @@ public class HeaderBasedSchemaResolutionResolver implements SchemaResolutionReso
 
     private int readSchemaId(ByteBuffer buffer) {
         if (buffer.remaining() < HEADER_SIZE) {
-            throw new SerializationException("Invalid payload size: " + buffer.remaining() + ", expected at least " + HEADER_SIZE);
+            throw new InvalidDataException("Invalid payload size: " + buffer.remaining() + ", expected at least " + HEADER_SIZE);
         }
 
         // Extract magic byte
         byte magicByte = buffer.get();
         if (magicByte != MAGIC_BYTE) {
-            throw new SerializationException("Unknown magic byte: " + magicByte);
+            throw new InvalidDataException("Unknown magic byte: " + magicByte);
         }
 
         // Extract schema ID
