+++ core/src/main/java/kafka/automq/table/transformer/RegistrySchemaAvroConverter.java
@@ -19,6 +19,7 @@
 
 package kafka.automq.table.transformer;
 
+import kafka.automq.table.process.exception.InvalidDataException;
 import kafka.automq.table.worker.convert.AvroToIcebergVisitor;
 import kafka.automq.table.worker.convert.IcebergRecordConverter;
